Update I2c modules to tolerance. Fix uint8_t to uint16_t tolerances.

+++ src/xpcc/architecture/platform/peripheral/spi/software/spi_master.hpp
@@ -46,7 +46,7 @@ public:
 public:
 	// start documentation inherited
 	template< class clockSource, uint32_t baudrate,
-			uint8_t tolerance = xpcc::Tolerance::FivePercent >
+			uint16_t tolerance = xpcc::Tolerance::FivePercent >
 	static inline void
 	initialize();
